debug

+++ ggpplayer/node.h
@@ -13,8 +13,8 @@ class Node
 {
 public:
 	Propositions state_;
-	int points_;
-	int attemps_;
+	long long points_;
+	long long attemps_;
 	vector<vector<Node > > sons_;
 	Node * parent_;
 	bool is_terminal_;
